+++ DisplayOptions.cs
@@ -13,7 +13,6 @@ namespace BoxOfYellow.ConsoleMarkdownRenderer
         /// Style applied to the expansion title that follows an
         /// <see cref="Markdig.Extensions.Abbreviations.AbbreviationInline"/> (e.g. the
         /// <c>HyperText Markup Language</c> portion of <c>HTML (HyperText Markup Language)</c>).
-        /// Only used when <see cref="ShowAbbreviationTitle"/> is <see langword="true"/>.
         /// </summary>
         public TextStyle AbbreviationTitle { get; set; } = new(decoration: TextDecoration.Dim);
 
@@ -54,16 +53,6 @@ namespace BoxOfYellow.ConsoleMarkdownRenderer
         /// Style applied to the term label of a <see cref="Markdig.Extensions.DefinitionLists.DefinitionTerm"/>.
         /// </summary>
         public TextStyle DefinitionTerm { get; set; } = new(decoration: TextDecoration.Bold);
-
-        /// <summary>
-        /// When <see langword="true"/> (the default), the expansion title of an
-        /// <see cref="Markdig.Extensions.Abbreviations.AbbreviationInline"/> is appended in
-        /// parentheses after the abbreviation text (e.g. <c>HTML (HyperText Markup Language)</c>),
-        /// styled with <see cref="AbbreviationTitle"/>.
-        /// When <see langword="false"/>, only the abbreviation text is rendered, keeping
-        /// the output compact.
-        /// </summary>
-        public bool ShowAbbreviationTitle { get; set; } = true;
 
         /// <summary>
         /// When set to true, the Info field from <see cref="Markdig.Syntax.FencedCodeBlock"/> (e.g., the language identifier) will be displayed.
@@ -231,7 +220,6 @@ namespace BoxOfYellow.ConsoleMarkdownRenderer
             MathBlockLabelText = this.MathBlockLabelText,
             MathInline = this.MathInline,
             QuotedBlock = this.QuotedBlock,
-            ShowAbbreviationTitle = this.ShowAbbreviationTitle,
             ShowFencedCodeBlockInfo = this.ShowFencedCodeBlockInfo,
             Strikethrough = this.Strikethrough,
             Subscript = this.Subscript,