+++ ObjectRenderers/ConsoleAbbreviationInlineRenderer.cs
@@ -7,14 +7,11 @@ namespace BoxOfYellow.ConsoleMarkdownRenderer.ObjectRenderers
     {
         protected override void Write(ConsoleRenderer renderer, AbbreviationInline obj)
         {
-            renderer.WriteEscape(obj.Abbreviation.Label);
-            if (renderer.Options.ShowAbbreviationTitle)
-            {
-                renderer
-                    .AddInLine($" ([{renderer.Options.AbbreviationTitle.ToSpectreStyle().ToMarkup()}]")
-                    .WriteEscape(obj.Abbreviation.Text.ToString())
-                    .AddInLine("[/])");
-            }
+            renderer
+                .WriteEscape(obj.Abbreviation.Label)
+                .AddInLine($" ([{renderer.Options.AbbreviationTitle.ToSpectreStyle().ToMarkup()}]")
+                .WriteEscape(obj.Abbreviation.Text.ToString())
+                .AddInLine("[/])");
         }
     }
 }